Extract component of sign transaction detail panel
Refatoring to prepare for signing with Ledger hardware.
Extraction of the new component `SignTransactionDetailPanel`
looks like a natural step to simplify `SignTransaction`.
